refactor(test): extract shared anvil/funding fixtures into @x402r/test-fixtures
