+++ examples/package.json
@@ -21,6 +21,7 @@
     "@x402/express": "2.14.0",
     "@x402/fetch": "2.14.0",
     "@x402r/evm": "0.2.0-alpha.1",
+    "@x402r/test-fixtures": "workspace:*",
     "express": "^5.0.0"
   }
 }
